requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
deepspeed==0.12.6
diffusers==0.29.1
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
easydict==1.13
egl_probe==1.0.2
einops==0.6.1
einops-exts==0.0.4
email_validator==2.2.0
evdev==1.7.1
exceptiongroup==1.2.1
Farama-Notifications==0.0.4
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
flash_attn==2.6.3
fonttools==4.51.0
fsspec==2024.5.0
ftfy==6.2.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
glfw==2.7.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.63.0
gymnasium==1.0.0
h11==0.14.0
h5py==3.11.0
hjson==3.1.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.23.4
hydra-core==1.3.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
llvmlite==0.41.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
# Editable install with no version control (metaworld==2.0.0)
-e git+https://github.com/NVlabs/mimicgen_environments.git@45db4b35a5a79e82ca8a70ce1321f855498ca82c#egg=mimicgen_envs
mpmath==1.3.0
mujoco==2.3.2
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.35.7
opencv-python==4.9.0.80
orjson==3.10.5
packaging==24.0
pandas==2.0.3
peft==0.11.1
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pluggy==1.5.0
protobuf==4.25.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
pynput==1.7.7
pynvml==11.5.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-xlib==0.33
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
-e git+https://github.com/ARISE-Initiative/robomimic.git@5dee58f9cc1235010d0877142b54d0e82dd23986#egg=robomimic
-e git+https://github.com/ARISE-Initiative/robosuite.git@eafb81f54ffc104f905ee48a16bb15f059176ad3#egg=robosuite
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.6.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.15.1
tomli==2.0.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.37.2
triton==2.2.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
wandb==0.17.2
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
zipp==3.18.2
